Refine the code
That's a pretty poor name for the change that's been made, but I needed
something succinct. So, here's a longer version of what changed. The
code for creating a link using the extracted URL path and template has
been refactored into a utility function. Secondly, the names of the
variables and functions have been cleaned up a little to make them as
meaningful and intuitive as possible. This was helped by writing some
code comments that helped me clarify and refine my thinking.